refactor(limits): fold provider labels into the provider catalog
The previous commit put the display names in a desktop-only module so that renaming a provider would not stale the Hub build marker. That reasoning does not survive the numbers: across this project's history 19 commits added, removed or reordered a provider id — each of which moves the marker regardless of where labels live — and exactly one renamed a label on its own. The split bought one avoided notice at the cost of a second file and a second UMD global.

So LIMIT_PROVIDER_CATALOG now carries id, label and settingsLabel together, with LIMIT_PROVIDER_IDS and LIMIT_PROVIDER_LABELS derived from it, and the renderer and the mac widget snapshot both read it. Loading a Hub-core module in the renderer is not new: index.html already loads currency, subscriptionDisplay, projectKey and reasonixSessionGuard the same way.

The embedded Hub is unaffected either way. It reports currentHubBuild() from the registry the app itself ships, so it compares against itself and always reads as current; the notice can only reach a separately deployed Node hub or Worker.
