+++ AGENTS.md
@@ -94,8 +94,8 @@ One caveat on top of the table:
 
 - Self-synced clients (cursor/antigravity) additionally go in `SELF_SYNCED_CLIENTS`; parse-local clients must NOT.
 - Targeted watch ticks make the client id a correctness surface, because the scan is keyed on it from two independent directions: `clientWatchCandidates()` decides which id a changed path maps to, and `normalizeClientName()` decides which id tokscale's rows land under. Three invariants keep them aligned — the id must be a fixed point of `normalizeClientName()` (so the partition a targeted scan writes is the one it cleared); every tokscale alias in `TOKSCALE_CLIENT_ALIASES` must normalize back to its parent id and be expanded by `tokscaleClientFilter()` (so targeting the parent still scans the alias, as with `antigravity` / `antigravity-cli`); and the filter must never emit `synthetic`. The first two are correctness: break either and a watch tick zeroes a client's partition, feeding a negative delta into month/allTime until the next full scan. The third is performance — `synthetic` makes tokscale enable *every* client, so the targeted scan silently degrades into a full one with correct numbers and none of the saving. Don't diagnose one as the other. `tests/shared/clientPartitionInvariants.test.js` enforces all three.
-- Limits providers declare their id and order once, in `LIMIT_PROVIDER_IDS` (`src/shared/limitProviders.js`), and their display name once, in `src/shared/limitProviderLabels.js`. The renderer's `LIMIT_PROVIDERS` and the mac widget snapshot both derive from those, so a new provider still needs account settings when applicable, tray artwork, a `.row-icon-<id>` mask, an explicit case in the Swift `WidgetFormat.provider` mirror, and every README table. `LIMIT_PROVIDER_IDS` defines the new-install order; a changed default must not overwrite a saved custom order.
-- Labels live outside `limitProviders.js` on purpose: that file is part of the portable Hub core hashed into `hubBuildRegistry.json`, so folding desktop-only presentation into it would make every renamed provider bump the core build id and tell self-hosted Hubs to redeploy for a change their runtime cannot observe. Same reason `limitProviderLabels.js` stays out of `WORKER_SHARED_MODULES`.
+- Limits providers declare their id, order and display name once, in `LIMIT_PROVIDER_CATALOG` (`src/shared/limitProviders.js`); `LIMIT_PROVIDER_IDS` and `LIMIT_PROVIDER_LABELS` are derived from it, and the renderer's `LIMIT_PROVIDERS` and the mac widget snapshot both read it. A new provider still needs account settings when applicable, tray artwork, a `.row-icon-<id>` mask, an explicit case in the Swift `WidgetFormat.provider` mirror, and every README table. The catalog order is the new-install order; a changed default must not overwrite a saved custom order.
+- That file is in the portable Hub core hashed into `hubBuildRegistry.json`, so editing it stales the Hub build marker. That is accepted rather than avoided: adding or reordering a provider changes the id list anyway, and a label has been renamed on its own once. Splitting the labels into a desktop-only module to dodge the marker was tried and reverted — it bought one avoided notice at the cost of a second file and a second global. Do not re-split it.
 
 ### Data flow contract
