+++ src/electron/renderer/index.html
@@ -1752,7 +1752,7 @@
       <span id="floatingBubbleContent" aria-hidden="true">Σ</span>
     </button>
     <script src="../../shared/clientCatalog.js"></script>
-    <script src="../../shared/limitProviderLabels.js"></script>
+    <script src="../../shared/limitProviders.js"></script>
     <script src="../../shared/limitBalanceDisplay.js"></script>
     <script src="trayBars.js"></script>
     <script src="usageCharts.js"></script>
