+++ worker/src/shared/hubBuildRegistry.json
@@ -113,6 +113,10 @@
       {
         "revision": 28,
         "buildId": "sha256:ff1bcb7b4b18e8478dba042482f3af3a8a0688853d64164403e411ace363ea58"
+      },
+      {
+        "revision": 29,
+        "buildId": "sha256:25bd090ffc832bcda78079e9577bccd837e0cad585014209dc0c8a6176f1d94c"
       }
     ],
     "node-hub": [
